refactor: extract MixedAudioTrackInsertionOffsets.swift from MixedAudioTypes.swift
Splits the `MixedAudioTrackInsertionOffsets` struct (with its CMTime
accessors) out of MixedAudioTypes.swift so that file holds only
`MixedAudioSourceStartTimes`.

Byte-identical move of lines 16-43. Both files import CoreMedia +
Foundation.

MixedAudioTypes.swift: 39 → 14 lines. Tests: 667.
